requirements.txt
requests
pandas
matplotlib
seaborn
python-dotenv
tqdm
adjustText
plotly